requirements.txt
django==3.2
djangorestframework
django-rest-framework
django-filter
drf-nested-routers
django-url-filter
django-cors-headers
psycopg2-binary
python-dotenv
whitenoise